update rule names
