+++ lib/rules/index.js
@@ -22,6 +22,6 @@ module.exports = new LazyLoadingRuleMap(Object.entries({
     "no-moment-js": () => require("./no-moment-js"),
     "no-date-fns": () => require("./no-date-fns"),
     "no-autoplay": () => require("./no-autoplay"),
-    "lighter-video-files": () => require("./lighter-video-files"),
-    "lighter-image-files": () => require("./lighter-image-files")
+    "lighter-video-formats": () => require("./lighter-video-formats"),
+    "lighter-image-formats": () => require("./lighter-image-formats")
 }));
