+++ lib/rules/lighter-image-formats.js
@@ -1,7 +1,7 @@
 /**
-* @fileoverview Rule to avoid heavy video file formats
-* @author Martyna Babiak
-*/
+ * @fileoverview Rule to avoid heavy image file formats
+ * @author Martyna Babiak
+ */
 
 "use strict";
 
@@ -21,30 +21,29 @@ module.exports = {
         type: "problem",
 
         docs: {
-            description: "Avoid heavy formats of video files",
+            description: "Avoid heavy formats of image files",
             category: "Code improvements",
             recommended: false,
-            url: "https://ec0lint.com/features/lighter-video-files"
+            url: "https://ec0lint.com/features/lighter-image-formats",
         },
         messages: {
-          rejected: 'Format of video files can be changed to WebM. Your video can be converted online at https://www.veed.io/convert/video-converter\n' +
- 		            'Estimated CO2 reduction that you can achieve by converting your file is: '
+            rejected:
+                "Format of image files can be changed to WebP or SVG. Your image can be converted online at https://cloudconvert.com/\n" +
+                "Estimated CO2 reduction that you can achieve by converting your file is: ",
         },
-        schema: [] // no options,
+        schema: [], // no options,
     },
-    create: function(context) {
-
+    create: function (context) {
         /**
-         * Returns if the video file format is heavy
-         * @param {string} name
+         * Returns if the image file format is heavy
+         * @param {string} name Module name
          * @returns {boolean} Returns true or false
          */
         function isBanned(name) {
-
             if (!name) {
                 return false;
             }
-            const result = name.match(/\.ogg|\.mp4|\.m4a|\.m4p|\.m4b|\.m4r|\.m4v/);
+            const result = name.match(/\.png|\.ppm|\.ps|\.rgb/);
 
             return result;
         }
@@ -56,15 +55,16 @@ module.exports = {
          * @returns {void}
          */
         function reportMessage(node, name) {
-
             if (isBanned(name)) {
-                context.report({ node, messageId: "rejected", data: { name }});
+                context.report({ node, messageId: "rejected", data: { name } });
             }
         }
 
-        return moduleVisitor.default((source, node) => {
-            reportMessage(node, source.value);
-        },
-        { commonjs: true });
-    }
+        return moduleVisitor.default(
+            (source, node) => {
+                reportMessage(node, source.value);
+            },
+            { commonjs: true }
+        );
+    },
 };
