+++ lib/rules/lighter-video-formats.js
@@ -1,7 +1,7 @@
 /**
- * @fileoverview Rule to avoid heavy image file formats
- * @author Martyna Babiak
- */
+* @fileoverview Rule to avoid heavy video file formats
+* @author Martyna Babiak
+*/
 
 "use strict";
 
@@ -21,29 +21,30 @@ module.exports = {
         type: "problem",
 
         docs: {
-            description: "Avoid heavy formats of image files",
+            description: "Avoid heavy formats of video files",
             category: "Code improvements",
             recommended: false,
-            url: "https://ec0lint.com/features/lighter-image-files",
+            url: "https://ec0lint.com/features/lighter-video-formats"
         },
         messages: {
-            rejected:
-                "Format of image files can be changed to WebP or SVG. Your image can be converted online at https://cloudconvert.com/\n" +
-                "Estimated CO2 reduction that you can achieve by converting your file is: ",
+          rejected: 'Format of video files can be changed to WebM. Your video can be converted online at https://www.veed.io/convert/video-converter\n' +
+ 		            'Estimated CO2 reduction that you can achieve by converting your file is: '
         },
-        schema: [], // no options,
+        schema: [] // no options,
     },
-    create: function (context) {
+    create: function(context) {
+
         /**
-         * Returns if the image file format is heavy
-         * @param {string} name Module name
+         * Returns if the video file format is heavy
+         * @param {string} name
          * @returns {boolean} Returns true or false
          */
         function isBanned(name) {
+
             if (!name) {
                 return false;
             }
-            const result = name.match(/\.png|\.ppm|\.ps|\.rgb/);
+            const result = name.match(/\.ogg|\.mp4|\.m4a|\.m4p|\.m4b|\.m4r|\.m4v/);
 
             return result;
         }
@@ -55,16 +56,15 @@ module.exports = {
          * @returns {void}
          */
         function reportMessage(node, name) {
+
             if (isBanned(name)) {
-                context.report({ node, messageId: "rejected", data: { name } });
+                context.report({ node, messageId: "rejected", data: { name }});
             }
         }
 
-        return moduleVisitor.default(
-            (source, node) => {
-                reportMessage(node, source.value);
-            },
-            { commonjs: true }
-        );
-    },
+        return moduleVisitor.default((source, node) => {
+            reportMessage(node, source.value);
+        },
+        { commonjs: true });
+    }
 };
