+++ packages/ec0lint-config-ec0lint/default.yml
@@ -5,8 +5,8 @@ extends:
 rules:
    avoid-lodash: "error"
    lighter-http: "error"
-   lighter-image-files": "error"
-   lighter-video-files": "error"
+   lighter-image-formats": "error"
+   lighter-video-formats": "error"
    no-ajax-events: "error"
    no-ajax: "error"
    no-animate: "error"
