+++ tools/rule-types.json
@@ -10,7 +10,7 @@
     "no-box-model": "problem",
     "no-date-fns": "problem",
     "no-autoplay": "problem",
-    "lighter-image-files": "problem",
-    "lighter-video-files": "problem",
+    "lighter-image-formats": "problem",
+    "lighter-video-formats": "problem",
     "no-moment-js": "problem"
 }
